Move request handler in common folder
