+++ src/request_handler/device_description_handler.cc
